+++ ChangeLog.md
@@ -1,5 +1,6 @@
 ## 0.0.23
 - Fixed case of commas in Map keys
+- Fixed keystore/truststore-pw typo (was pwd)
 
 ## 0.0.22
 - Added support for gzipped input files